Use comm_md5_prefix as fake mmap's filename_md5_prefix for kernel sample's main mapping.
When a kernel sample don't have a main mapping, a fake main mmap is created with empty filename_md5_prefix. Thus the process name is always empty for samples with empty comm name. In order to assign such samples with the correct process names (aka execname), use the comm_md5_prefix of the comm event (whose pid maps to the sample.pid) as the fake mmap's filename_md5_prefix.

PiperOrigin-RevId: 279122865