Preserve the full mapping name for the main kernel image.
The mmap events for vmlinux include a DSO with a relocation symbol suffix (the "_stext" in
"[kernel.kallsyms]_stext") that's necessary for correct symbolizaiton.

They are differemt from the DSO names in build-id events (only "[kernel.kallsyms]").

Instead of renaming the former to the latter (in order to match mmaps to build-ids),
preserve the full mmap DSO names, and generate alternate build-id events with names
that include the suffix.

PiperOrigin-RevId: 432226042