Distinguish between `cache_latency` and `total_latency` coming from x86 machines with WEIGHT or WEIGHT_STRUCT.
Both newer Intel machines and AMD machines with IBS populate both fields, so we propagate them accordingly as Profile proto labels.
Older Intel machines only populated the `weight` field, with the "total" issue-to-retirement latency, but we mistakenly used to convert that to `cache_latency`, now
it is in `total_latency`.

PiperOrigin-RevId: 641530291