.bazelrc: Configure to not emit warnings from 3rd party code.
Code that we import via MODULE.bazel generates warnings that
are not really actionable from this projects' perspective.
Use new bazel6+ feature to set different options for files matching a
regex, in this case all external/ files shall have switched-off warnings.

Also: explicitly give the -xc++ option for c++ files. Bazel does
not invoke `clang++` but the generic `clang` driver for all files and
hopes that it auto-detects the file. This is not true for all
clang installations, and in particular clangd, clang-tidy etc. have
issues not knowing the type of file.
PiperOrigin-RevId: 693428479