Speed up quipper AddressMapper when having a large number of mmaps.
AddressMapper is a performance bottleneck when processing perf data with large number of mmap events, because the current implementation scans the existing mappings to find a place for new mapping. This change uses a heuristic to try to add a new range at the end of the list when the size of the list passing certain threshold. Essentially, the time complexity will change from O(N^2) to O(N*log(N)) when having a large number of mmaps (N >= 10K).

This CL also adds the test to make sure that when having a large number of mmaps, the parsing can finish in a reasonable time.

PiperOrigin-RevId: 786751094